[Feat] #15 - 스터디 모집 & 스터디 Command API 구현

+++ src/main/java/com/threego/algo/studyrecruit/query/dto/StudyRecruitDetailDTO.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class StudyRecruitDetailDTO {
 
-    private Long id;                    // 모집글 ID
+    private int id;                    // 모집글 ID
     private String title;               // 제목
     private String content;             // 내용
     private String memberNickname;      // 작성자 닉네임
@@ -21,9 +21,9 @@ public class StudyRecruitDetailDTO {
     private String startDate;           // 스터디 시작일
     private String endDate;             // 스터디 종료일
     private String expiresAt;           // 모집 마감일
-    private Integer capacity;           // 모집 정원
-    private Integer participantCount;   // 현재 참여자 수
-    private Integer commentCount;       // 댓글 수
+    private int capacity;           // 모집 정원
+    private int participantCount;   // 현재 참여자 수
+    private int commentCount;       // 댓글 수
     private String createdAt;           // 작성일
     private String updatedAt;           // 수정일
 }
